refactor: replace `setupTauriCloseHandlers` with `useTauriCloseHandler` hook
- Migrate Tauri app close handling to a new React hook for better modularity and integration with app lifecycle.
- Remove `tauriLifecycle` utilities from SDK, simplifying the API surface.
- Update macOS/Windows window state save logic before hiding to tray.
